requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
bandit==1.7.5
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.6
django-environ==0.11.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gitdb==4.0.11
GitPython==3.1.40
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
openapi-codec==1.3.2
packaging==23.2
pbr==5.11.1
prompt-toolkit==3.0.39
psycopg2==2.9.9
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
rich==13.6.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
stevedore==5.1.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
